requirements.txt
pydantic==2.8.2
PyYAML==6.0.2
numpy==1.26.4
pandas==2.2.2
websockets==12.0
requests==2.32.3
redis==5.0.7
prometheus-client==0.20.0
streamlit==1.38.0
matplotlib==3.8.4
typing-extensions==4.12.2
